break apart instance destroy method into actions and modals
