+++ troposphere/static/js/modals/InstanceModals.js
@@ -1,7 +1,8 @@
 define(function (require) {
 
   return {
-    createAndAddToProject: require('./instance/createAndAddToProject').createAndAddToProject
+    createAndAddToProject: require('./instance/createAndAddToProject').createAndAddToProject,
+    destroy: require('./instance/destroy').destroy
   };
 
 });
